hopefully fix windows build

+++ src/test/test_evtbin.cxx
@@ -450,7 +450,7 @@ void EvtBinTest::testHealpixBinner() {
   for(int order=0;order!=13;order++){
     HealpixBinner binner("NESTED", order, true, "Binner");
     long nbins= binner.getNumBins();
-    long int nside = pow((long int)2,order);
+    long int nside = pow((long double)2,order);
     long true_nbins=12*nside*nside;
     if(nbins!=true_nbins) {
       m_failed=true;
